requirements.txt
streamlit==1.3.0
pandas==1.3.5
scikit-learn==0.24.2
Pillow==8.4.0
